+++ src/main/java/de/dm/intellij/liferay/language/resourcebundle/ResourceBundleReferenceProvider.java
@@ -1,0 +1,23 @@
+package de.dm.intellij.liferay.language.resourcebundle;
+
+import com.intellij.lang.properties.ResourceBundleReference;
+import com.intellij.lang.properties.psi.PropertiesFile;
+import com.intellij.lang.properties.references.I18nUtil;
+import com.intellij.lang.properties.references.PropertyReference;
+import com.intellij.psi.PsiElement;
+import com.intellij.psi.PsiReference;
+import com.intellij.psi.PsiReferenceProvider;
+import com.intellij.util.ProcessingContext;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.List;
+
+public class ResourceBundleReferenceProvider extends PsiReferenceProvider {
+
+    @Override
+    public PsiReference @NotNull [] getReferencesByElement(@NotNull PsiElement element, @NotNull ProcessingContext context) {
+        return new PsiReference[] {
+                new CustomResourceBundleReference(element)
+        };
+    }
+}