+++ src/test/java/de/dm/intellij/liferay/language/resourcebundle/ResourceBundleReferenceProviderTest.java
@@ -1,0 +1,36 @@
+package de.dm.intellij.liferay.language.resourcebundle;
+
+import com.intellij.codeInsight.completion.CompletionType;
+import com.intellij.testFramework.LightProjectDescriptor;
+import com.intellij.testFramework.fixtures.LightJavaCodeInsightFixtureTestCase;
+import de.dm.intellij.test.helper.LightProjectDescriptorBuilder;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.List;
+
+public class ResourceBundleReferenceProviderTest extends LightJavaCodeInsightFixtureTestCase {
+
+    private static final String TEST_DATA_PATH = "testdata/de/dm/intellij/liferay/language/resourcebundle/ResourceBundleReferenceProviderTest";
+
+    @NotNull
+    @Override
+    protected LightProjectDescriptor getProjectDescriptor() {
+        return LightProjectDescriptorBuilder.DEFAULT_PROJECT_DESCRIPTOR;
+    }
+
+    @Override
+    protected String getTestDataPath() {
+        return TEST_DATA_PATH;
+    }
+
+    public void testResourceBundleCodeCompletion() {
+        myFixture.configureByFiles("MyResourceBundleLoader.java", "content/Language.properties", "content/Language_en.properties");
+        myFixture.complete(CompletionType.BASIC, 1);
+
+        List<String> strings = myFixture.getLookupElementStrings();
+
+        assertTrue(strings.contains("content.Language"));
+        assertTrue(strings.contains("content.Language_en"));
+    }
+
+}