+++ changelog.md
@@ -4,6 +4,7 @@ Liferay Plugin for IntelliJ
 All notable changes to this project will be documented in this file.
 
 ## 0.0.14
+- Feature: Resolve Resource Bundles
 
 ## 0.0.13
 - Feature: Language injections for script and template tags in Liferay workflow definition files